% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpos.R
\name{findpos}
\alias{findpos}
\title{\code{Find SNP Position}}
\usage{
findpos(blastbin = NULL, db, fa, iden, len)
}
\arguments{
\item{blastbin}{The path of binary code for NCBI BLAST program. [String, default=NULL]}

\item{db}{Path of database to BLAST against. [String]}

\item{fa}{A fasta file of the sequences of interest. [String]}

\item{iden}{Percent identity of the alignment. [interger, default=95]}

\item{len}{Minimum length of the alignment. [interger, default=100]}
}
\value{
return a data.frame.
}
\description{
Find the physical position of a short sequence in a given genome.
}
\examples{

library("Biostrings")
library("findpos")
### Find the help document for the function
?findpos
### run
 res <- findpos(blastbin = "~/bin/ncbi-blast-2.4.0+/bin",
                db="example/16SMicrobialDB/16SMicrobial",
                fa="example/rna.fasta", iden=95, len=100)

}

