% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLAST.R
\name{getblast.BLAST}
\alias{getblast.BLAST}
\title{\code{Get BLAST Results in a table}}
\usage{
getblast.BLAST(object, seq, BLAST_args = "", custom_format = "", ...)
}
\arguments{
\item{object}{BLAST object. [BLAST]}

\item{seq}{Biostring object. [Biostrings]}
}
\value{
return a data.frame.
}
\description{
format the blast results.
}
\examples{
inputdf <- data.frame(fq1="fq_1.fq", fq2="f1_2.fq", out="mysample",
                 group="g1", sample="s1", PL="illumina", LB="lib1", PU="unit1")

run_GATK(inputdf,
         ref.fa="~/dbcenter/Ecoli/reference/Ecoli_k12_MG1655.fasta",
         gatkpwd="$HOME/bin/GenomeAnalysisTK-3.5/GenomeAnalysisTK.jar",
         picardpwd="$HOME/bin/picard-tools-2.1.1/picard.jar",
         markDup=TRUE,
         realignInDels=FALSE, indels.vcf="indels.vcf",
         recalBases=FALSE, dbsnp.vcf="dbsnp.vcf",
         email=NULL, runinfo = c(FALSE, "bigmemh", 1))

}

