% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLAST.R
\name{getblast}
\alias{getblast}
\title{\code{Get BLAST Results in a table}}
\usage{
getblast(object, seq, BLAST_args = "", custom_format = "", ...)
}
\arguments{
\item{object}{BLAST object. [BLAST]}

\item{seq}{Biostring object. [Biostrings]}
}
\value{
return a data.frame.
}
\description{
format the blast results.
}
\examples{

}

