% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpos.R
\name{findpos}
\alias{findpos}
\title{\code{Find SNP Position}}
\usage{
findpos(blastbin = NULL, db, fa, iden, len)
}
\arguments{
\item{blastbin}{The path of binary code for NCBI BLAST program. [String, default=NULL]}

\item{db}{Path of database to BLAST against. [String]}

\item{fa}{A fasta file of the sequences of interest. [String]}

\item{iden}{Percent identity of the alignment. [interger, default=95]}

\item{len}{Minimum length of the alignment. [interger, default=100]}
}
\value{
return a data.frame.
}
\description{
Find the physical position of a short sequence in a given genome.
}
\examples{
inputdf <- data.frame(fq1="fq_1.fq", fq2="f1_2.fq", out="mysample",
                 group="g1", sample="s1", PL="illumina", LB="lib1", PU="unit1")

run_GATK(inputdf,
         ref.fa="~/dbcenter/Ecoli/reference/Ecoli_k12_MG1655.fasta",
         gatkpwd="$HOME/bin/GenomeAnalysisTK-3.5/GenomeAnalysisTK.jar",
         picardpwd="$HOME/bin/picard-tools-2.1.1/picard.jar",
         markDup=TRUE,
         realignInDels=FALSE, indels.vcf="indels.vcf",
         recalBases=FALSE, dbsnp.vcf="dbsnp.vcf",
         email=NULL, runinfo = c(FALSE, "bigmemh", 1))

}

